requirements.txt
tensorflow-gpu
albumentations
matplotlib
pykitti
tqdm
numpy
keras
requests